+++ Libraries/Components/TextInput/TextInputState.js
@@ -73,7 +73,7 @@ function blurField(textFieldID: ?number) {
 /**
  * @param {number} TextInputID id of the text field to focus
  * Focuses the specified text field
- * noop if the text field was already focused
+ * noop if the text field was already focused or if the field is not editable
  */
 function focusTextInput(textField: ?ComponentRef) {
   if (typeof textField === 'number') {
@@ -86,7 +86,15 @@ function focusTextInput(textField: ?ComponentRef) {
     return;
   }
 
-  if (currentlyFocusedInputRef !== textField && textField != null) {
+  if (textField != null) {
+    const fieldCanBeFocused =
+      currentlyFocusedInputRef !== textField &&
+      // $FlowFixMe - `currentProps` is missing in `NativeMethods`
+      textField.currentProps?.editable !== false;
+
+    if (!fieldCanBeFocused) {
+      return;
+    }
     focusInput(textField);
     if (Platform.OS === 'ios') {
       // This isn't necessarily a single line text input