+++ Libraries/TypeSafety/RCTConvertHelpers.mm
@@ -14,7 +14,7 @@ bool RCTBridgingToBool(id value)
   return [RCTConvert BOOL:value] ? true : false;
 }
 
-folly::Optional<bool> RCTBridgingToOptionalBool(id value)
+std::optional<bool> RCTBridgingToOptionalBool(id value)
 {
   if (!RCTNilIfNull(value)) {
     return {};
@@ -33,7 +33,7 @@ NSString *RCTBridgingToOptionalString(id value)
 }
 
 
-folly::Optional<double> RCTBridgingToOptionalDouble(id value)
+std::optional<double> RCTBridgingToOptionalDouble(id value)
 {
   if (!RCTNilIfNull(value)) {
     return {};