+++ packages/react-native-codegen/src/generators/modules/GenerateModuleJniH.js
@@ -103,6 +103,56 @@ include $(BUILD_SHARED_LIBRARY)
 `;
 };
 
+// Note: this CMakeLists.txt template includes dependencies for both NativeModule and components.
+const CMakeListsTemplate = ({
+  libraryName,
+}: $ReadOnly<{libraryName: string}>) => {
+  return `# Copyright (c) Meta Platforms, Inc. and affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+cmake_minimum_required(VERSION 3.13)
+set(CMAKE_VERBOSE_MAKEFILE on)
+
+file(GLOB react_codegen_SRCS CONFIGURE_DEPENDS *.cpp react/renderer/components/${libraryName}/*.cpp)
+
+add_library(
+  react_codegen_${libraryName}
+  SHARED
+  \${react_codegen_SRCS}
+)
+
+target_include_directories(react_codegen_${libraryName} PUBLIC . react/renderer/components/${libraryName})
+
+target_link_libraries(
+  react_codegen_${libraryName}
+  fbjni
+  folly_runtime
+  glog
+  ${libraryName !== 'rncore' ? 'react_codegen_rncore' : ''}
+  react_debug
+  react_nativemodule_core
+  react_render_core
+  react_render_debug
+  react_render_graphics
+  rrc_view
+  turbomodulejsijni
+  yoga
+)
+
+target_compile_options(
+  react_codegen_${libraryName}
+  PRIVATE
+  -DLOG_TAG=\\"ReactNative\\"
+  -fexceptions
+  -frtti
+  -std=c++17
+  -Wall
+)
+`;
+};
+
 module.exports = {
   generate(
     libraryName: string,
@@ -136,6 +186,7 @@ module.exports = {
           libraryName: libraryName,
         }),
       ],
+      ['jni/CMakeLists.txt', CMakeListsTemplate({libraryName: libraryName})],
     ]);
   },
 };